codegen: move return wrapper getter into one function
